fix: Resolve Flutter analysis errors
- Fix Icon naming errors (remove _rounded suffix)
- Fix chart viewer formatting and type errors
- Remove .toInt() from double volume parameter
- Fix line ending issues in chart_viewer_screen.dart
- Add missing performance_chart.dart widget
- Fix const Icon errors in settings_screen.dart

All analysis errors resolved:
- Icon getters: Fixed 15+ icon name errors
- Type errors: Fixed int->double conversion
- Formatting: Fixed compressed parameter formatting
- Missing files: Added performance_chart.dart

+++ lib/screens/chart_viewer_screen.dart
@@ -321,7 +321,8 @@ class _ChartViewerScreenState extends State<ChartViewerScreen> {
         open: 290 + (index * 2.5),
         high: 298 + (index * 2.5),
         low: 287 + (index * 2.5),
-        close: 295 + (index * 2.5),volume: 1000000 + (index * 50000).toInt(),
+        close: 295 + (index * 2.5),
+        volume: (1000000 + (index * 50000)),
         time:
 DateTime.now().subtract(Duration(minutes: 50 - index)),
       );
